# Add specific 'defined?(@ivar)' instruction