Currently, the `defined?` operation is implemented with single, general VM instruction covering different cases of what's being checked. This PR creates a specific VM instruction for when the argument to `defined?` is an instance variable.

This allows us to eliminate a `putnil` instruction for that case, and the implementation is more straightforward. Additionally, we can make use of shape information in CRuby.

I wrote a very silly benchmark:

```ruby
require 'benchmark'

class IVarClass
  def set_ivar
    @foo = 1
  end

  def test_defined
    50000000.times { defined?(@foo) }
  end
end

obj = IVarClass.new
puts Benchmark.measure { obj.test_defined }

obj.set_ivar
puts Benchmark.measure { obj.test_defined }
```

I've run it a few times on `master` and this branch, with and without YJIT enabled. (I'm on a laptop, so runtimes may fluctuate a bit, so this is not exact science.)

Representative values on `master`:
```
$ ./ruby --yjit ivar_microbenchmark.rb
  3.580647   0.006810   3.587457 (  3.590800)
  3.516924   0.004903   3.521827 (  3.525773)
$ ./ruby ivar_microbenchmark.rb
  4.016235   0.008735   4.024970 (  4.029170)
  4.032165   0.006480   4.038645 (  4.041866)
```

On this branch:
```
$ ./ruby --yjit ivar_microbenchmark.rb
  3.172839   0.006284   3.179123 (  3.182953)
  3.125201   0.006086   3.131287 (  3.133984)
$ ./ruby ivar_microbenchmark.rb 
  3.447063   0.007094   3.454157 (  3.457770)
  3.490655   0.007111   3.497766 (  3.501535)
```

For these runs, it gives about 13% better performance on CRuby and about 11% better performance with YJIT. As mentioned, these numbers are not super stable on my laptop, but I'm consistently getting better than 10% both on CRuby and YJIT.

## Notes
For the CRuby version, I've added an extra parameter to `vm_getivar` so that I can re-use the logic. I'm not super proud of this, as it makes an already complex function a bit more complex. I've tried to be as gentle as possible. If somebody has other ideas to get the same effect in a cleaner way, let me know.

Also, this _will_ introduce a new internal instruction. I don't know how people feel about that.

Thanks a lot to @jhawthorn who has helped from start to finish with this PR!