@maximecb 2023-03-03
I feel like this is mostly an optimization for the interpreter. No reason the JIT can't use shapes to optimize `defined` already, it just hasn't been done. However, not strongly opposed to this. It's not a bad thing to make the interpreter faster, so I approved it.