@k0kubun 2023-03-03
How much impact did it have on non-microbenchmarks?