@maximecb 2023-03-03
> I think this is a very good idea for the interpreter. I'm not a fan of people doing `@foo = blah unless defined?(@foo)`, but it's something people do in the wild and we may as well have an IC for that case.

Good point Aaron 👍