@k0kubun 2023-03-03
[yjit-bench](https://github.com/shopify/yjit-bench) `--category headline` results:

## Interpreter
```
before: ruby 3.3.0dev (2023-02-27T11:07:19Z master 0aa50a03b1) [x86_64-linux]
after: ruby 3.3.0dev (2023-03-02T12:07:49Z add-defined_ivar-i.. 55a5156fc0) [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
activerecord   114.7        0.3         112.2       0.6         1.02          1.01
hexapdf        1809.9       0.7         1756.3      0.7         1.03          1.02
liquid-c       44.4         1.9         44.7        1.9         0.99          0.99
liquid-render  109.3        1.3         109.6       1.5         1.00          0.99
mail           94.8         0.3         93.3        0.6         1.02          1.00
psych-load     1453.5       0.2         1422.2      0.1         1.02          1.02
railsbench     1530.3       0.9         1522.3      1.0         1.01          1.01
ruby-lsp       47.1         12.8        46.3        13.1        1.02          0.99
sequel         90.1         1.8         90.4        1.7         1.00          1.00
-------------  -----------  ----------  ----------  ----------  ------------  -------------
```

## YJIT
```
before: ruby 3.3.0dev (2023-02-27T11:07:19Z master 0aa50a03b1) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-03-02T12:07:49Z add-defined_ivar-i.. 55a5156fc0) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
activerecord   67.4         0.5         67.5        0.3         1.00          1.00
hexapdf        1213.3       1.2         1204.9      1.4         1.01          1.00
liquid-c       32.0         3.2         32.2        2.9         0.99          1.00
liquid-render  62.5         2.9         62.4        3.3         1.00          0.99
mail           72.7         0.2         72.1        0.3         1.01          1.00
psych-load     1022.7       0.2         1014.9      0.1         1.01          1.00
railsbench     1004.8       1.3         1004.2      1.7         1.00          1.00
ruby-lsp       37.0         28.0        36.7        29.6        1.01          1.01
sequel         80.1         2.5         80.1        2.3         1.00          1.01
-------------  -----------  ----------  ----------  ----------  ------------  -------------
```

---

The interpreter seems 2~3% faster on some benchmarks (I can stably reproduce it). YJIT seems to be the same.