@maximecb 2023-03-08
@jhawthorn @olefriis Aaron @tenderlove has shown interest in implementing a faster version of this instruction for YJIT which takes advantage of shapes.

Would you be willing to pair with him on this? :)